Refactor files in `src/`
