+++ src/EndPoint/Api/Shared/NotFoundMiddleware.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Http;
+namespace App\EndPoint\Api\Shared;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
