+++ src/EndPoint/Console/Hello/Action.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Command;
+namespace App\EndPoint\Console\Hello;
 
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
@@ -10,8 +10,11 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Yiisoft\Yii\Console\ExitCode;
 
-#[AsCommand(name: 'hello', description: 'An example command', hidden: false)]
-final class HelloCommand extends Command
+#[AsCommand(
+    name: 'hello',
+    description: 'An example command',
+)]
+final class Action extends Command
 {
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
