+++ src/Shared/ApplicationParams.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App;
+namespace App\Shared;
 
 final readonly class ApplicationParams
 {
